+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2024-12-26
+
+- Add `Id` parameter to `Get-WsusComputerGroup` cmdlet
+- Rename `ComputerTargetGroup` to `TargetGroup` parameter in `Get-WsusComputerGroup` cmdlet
+
 ## 2024-12-21
 
 Initial commit